+++ client/src/extension.ts
@@ -93,6 +93,7 @@ export function activate(context: vscode.ExtensionContext) {
 	client.start()		// also starts server
 
 	context.subscriptions.push(vscode.commands.registerCommand("rpw65.renumberLocals", renumberCmd))
+	context.subscriptions.push(vscode.commands.registerCommand("rpw65.stepForward", stepForwardCmd ))
 	context.subscriptions.push(vscode.commands.registerCommand("rpw65.tabIndent", () => { cmd.tabIndentCmd(false) }))
 	context.subscriptions.push(vscode.commands.registerCommand("rpw65.tabOutdent", () => { cmd.tabIndentCmd(true) }))
 	context.subscriptions.push(vscode.commands.registerCommand("rpw65.delIndent", cmd.delIndentCmd))
@@ -288,4 +289,9 @@ function updateStatusItem() {
   }
 }
 
-//------------------------------------------------------------------------------
+export async function stepForwardCmd() {
+	await client.sendRequest(vsclnt.ExecuteCommandRequest.type, {
+		command: "rpw65.debugger",
+		arguments: ["stepForward"]
+	})
+}